+++ heimdallm/bifrosts/sql/sqlite/tests/select/test_evil.py
@@ -1,8 +1,8 @@
 import pytest
 
 from heimdallm.bifrosts.sql import exc
-from heimdallm.bifrosts.sql.sqlite.select.bifrost import SQLBifrost
-from heimdallm.bifrosts.sql.utils import FqColumn
+from heimdallm.bifrosts.sql.common import FqColumn
+from heimdallm.bifrosts.sql.sqlite.select.bifrost import Bifrost
 
 from .utils import PermissiveConstraints
 
@@ -16,7 +16,7 @@ from t1
 left join t2 on t1.jid = t2.jid
     """
 
-    bifrost = SQLBifrost.mocked(PermissiveConstraints())
+    bifrost = Bifrost.mocked(PermissiveConstraints())
     with pytest.raises(exc.IllegalJoinType) as e:
         bifrost.traverse(query)
     assert e.value.join_type == "OUTER_JOIN"
@@ -29,7 +29,7 @@ def test_no_select_star():
         def select_column_allowed(self, column: FqColumn) -> bool:
             return True
 
-    bifrost = SQLBifrost.mocked(MyConstraints())
+    bifrost = Bifrost.mocked(MyConstraints())
     with pytest.raises(exc.IllegalSelectedColumn) as e:
         bifrost.traverse(query)
 
@@ -42,7 +42,7 @@ def test_allow_count_star():
             # count(*) doesn't count as a column that needs to be allowlisted
             return False
 
-    bifrost = SQLBifrost.mocked(MyConstraints())
+    bifrost = Bifrost.mocked(MyConstraints())
 
     query = "select count(*) from t1"
     bifrost.traverse(query)
@@ -57,7 +57,7 @@ select t1.secret
 from t1 group
     """
 
-    bifrost = SQLBifrost.mocked(PermissiveConstraints())
+    bifrost = Bifrost.mocked(PermissiveConstraints())
     with pytest.raises(exc.ReservedKeyword) as e:
         bifrost.traverse(query)
     assert e.value.keyword == "group"
@@ -76,7 +76,7 @@ select t1.secret group
 from t1
     """
 
-    bifrost = SQLBifrost.mocked(PermissiveConstraints())
+    bifrost = Bifrost.mocked(PermissiveConstraints())
     with pytest.raises(exc.ReservedKeyword) as e:
         bifrost.traverse(query)
     assert e.value.keyword == "group"
@@ -95,5 +95,5 @@ select t1.col
 from t1
 where t1.col='let''s go'
     """
-    bifrost = SQLBifrost.mocked(PermissiveConstraints())
+    bifrost = Bifrost.mocked(PermissiveConstraints())
     bifrost.traverse(query)