+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.1.5 - 7/9/23
+
+- Refactor sqlite helpers to be general
+- Bugfix where disabling reconstruction resulted in an error
+- Allow counting disallowed columns
+- Bugfix api key not set when using openai completion method
+
 ## 0.1.4 - 7/5/23
 
 - Use an optional int instead of `math.inf` for an unrestricted SQL limit